Reworked the scene system to be more C++ style
Scene system is now a polymorphic class so that inputs are not globally
checked, but locally to the scene it self. This mean less complexity and
les CPU cycles wasted.

Added a special handling for the svcSleepThread function that fails to
be linked inside a virtual method, don't know why.

Bug: #1

Change-Id: I26ef85b254ec5e5773d81b4c4573670839d9ca94